+++ src/Enjin.Platform.Sdk/Enjin.Platform.Sdk.Marketplace/Schema/Fragments/FixedPriceDataFragment.cs
@@ -21,6 +21,11 @@ public class FixedPriceDataFragment : ListingDataFragment<FixedPriceDataFragment
     /// </summary>
     /// <param name="isIncluded">Whether the field is included.</param>
     /// <returns>This fragment for chaining.</returns>
+    /// <remarks>
+    /// This field ought to be requested when using this fragment type for a
+    /// <see cref="ListingDataFragment{TFragment}"/> or else the <see cref="ListingDataJsonConverter"/> may not be able
+    /// to determine the type.
+    /// </remarks>
     public FixedPriceDataFragment WithType(bool isIncluded = true)
     {
         return WithField("type", isIncluded);